add setup.py, put __main__.py stuff into a function
setup.py was written out so the module can be properly packaged.

the main stuff is now in a main function.  This is mostly to interact
with setuptools.  Note the setup.py entry_points member.  This specifies
this function so that if you build and then install this wheel, it will
also install a script on your system that will allow the user to just
run `charactergenerator` instead of `python -mcharactergenerator`.

If you set up a GUI for this, you'll change `console_scripts` to
`gui_scripts`, and it will interact cleanly with Windows python
packagers as well, automatically running your program as a GUI just from
clicking the generated .exe.